# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Anonymous Messaging Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  /**
   * @minLength 3
   * @maxLength 30
   */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginRequest {
  usernameOrEmail: string;
  password: string;
}

export interface UpdateAvatarRequest {
  avatarUrl: string;
}

export interface UserInfo {
  id: number;
  username: string;
  email: string;
  avatarUrl?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: UserInfo;
  token: string;
}

export interface Post {
  id: number;
  content: string;
  imageUrl?: string | null;
  isAnonymous: boolean;
  username?: string | null;
  avatarUrl?: string | null;
  createdAt: string;
}

export interface FeedResponse {
  posts: Post[];
  total: number;
  page: number;
  limit: number;
}

export interface CreatePostRequest {
  /**
   * @minLength 1
   * @maxLength 1000
   */
  content: string;
  imageUrl?: string | null;
  isAnonymous: boolean;
}

export interface UserProfile {
  id: number;
  username: string;
  avatarUrl?: string | null;
  createdAt: string;
}

export interface UserListResponse {
  users: UserProfile[];
}

export interface SendMessageRequest {
  /**
   * @minLength 1
   * @maxLength 1000
   */
  body: string;
  imageUrl?: string | null;
  guestSessionId?: string;
}

export interface Conversation {
  id: number;
  guestSessionId: string;
  anonymousAlias: string;
  lastMessage?: string | null;
  lastMessageAt?: string | null;
  unreadCount: number;
  createdAt: string;
}

export interface InboxResponse {
  conversations: Conversation[];
}

export interface InboxMessage {
  id: number;
  conversationId: number;
  body: string;
  imageUrl?: string | null;
  isRead: boolean;
  isFromOwner: boolean;
  createdAt: string;
}

export interface ConversationResponse {
  conversation: Conversation;
  messages: InboxMessage[];
}

export interface ReplyRequest {
  /**
   * @minLength 1
   * @maxLength 1000
   */
  body: string;
  imageUrl?: string | null;
}

export interface UnreadCountResponse {
  count: number;
}

export interface SenderReplyRequest {
  /**
   * @minLength 1
   * @maxLength 1000
   */
  body: string;
  imageUrl?: string | null;
  guestSessionId: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type GetFeedParams = {
  page?: number;
  limit?: number;
};

export type GetSentConversationsParams = {
  guestSessionId: string;
};

export type GetSentConversationParams = {
  guestSessionId: string;
};