# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Anonymous Messaging Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username or email already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me/avatar:
    patch:
      operationId: updateAvatar
      tags: [auth]
      summary: Update current user's avatar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAvatarRequest"
      responses:
        "200":
          description: Avatar updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /feed:
    get:
      operationId: getFeed
      tags: [feed]
      summary: Get public feed posts
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedResponse"

    post:
      operationId: createPost
      tags: [feed]
      summary: Create a public post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /inbox:
    get:
      operationId: getInbox
      tags: [inbox]
      summary: Get inbox conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inbox/{conversationId}:
    get:
      operationId: getConversation
      tags: [inbox]
      summary: Get messages in a conversation
      parameters:
        - in: path
          name: conversationId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inbox/{conversationId}/reply:
    post:
      operationId: replyToConversation
      tags: [inbox]
      summary: Reply to an anonymous conversation
      parameters:
        - in: path
          name: conversationId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplyRequest"
      responses:
        "201":
          description: Reply sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inbox/unread-count:
    get:
      operationId: getUnreadCount
      tags: [inbox]
      summary: Get count of unread messages
      responses:
        "200":
          description: Unread count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnreadCountResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required:
        - username
        - email
        - password
    LoginRequest:
      type: object
      properties:
        usernameOrEmail:
          type: string
        password:
          type: string
      required:
        - usernameOrEmail
        - password
    UpdateAvatarRequest:
      type: object
      properties:
        avatarUrl:
          type: string
      required:
        - avatarUrl
    UserInfo:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
        - createdAt
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserInfo"
        token:
          type: string
      required:
        - user
        - token
    Post:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        isAnonymous:
          type: boolean
        username:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - isAnonymous
        - createdAt
    FeedResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - posts
        - total
        - page
        - limit
    CreatePostRequest:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000
        imageUrl:
          type: string
          nullable: true
        isAnonymous:
          type: boolean
      required:
        - content
        - isAnonymous
...
    Conversation:
      type: object
      properties:
        id:
          type: integer
        guestSessionId:
          type: string
        anonymousAlias:
          type: string
        lastMessage:
          type: string
          nullable: true
        lastMessageAt:
          type: string
          format: date-time
          nullable: true
        unreadCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - guestSessionId
        - anonymousAlias
        - unreadCount
        - createdAt
    InboxResponse:
      type: object
      properties:
        conversations:
          type: array
          items:
            $ref: "#/components/schemas/Conversation"
      required:
        - conversations
    InboxMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        body:
          type: string
        imageUrl:
          type: string
          nullable: true
        isRead:
          type: boolean
        isFromOwner:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - body
        - isRead
        - isFromOwner
        - createdAt
    ConversationResponse:
      type: object
      properties:
        conversation:
          $ref: "#/components/schemas/Conversation"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/InboxMessage"
      required:
        - conversation
        - messages
    ReplyRequest:
      type: object
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 1000
        imageUrl:
          type: string
          nullable: true
      required:
        - body
    UnreadCountResponse:
      type: object
      properties:
        count:
          type: integer
      required:
        - count
...