# lib/api-zod/src/generated/types/createPostRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Anonymous Messaging Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePostRequest {
  /**
   * @minLength 1
   * @maxLength 1000
   */
  content: string;
  imageUrl?: string | null;
  isAnonymous: boolean;
}