# lib/api-zod/src/generated/types/inboxMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Anonymous Messaging Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface InboxMessage {
  id: number;
  conversationId: number;
  body: string;
  imageUrl?: string | null;
  isRead: boolean;
  isFromOwner: boolean;
  createdAt: Date;
}