# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Anonymous Messaging Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Post {
  id: number;
  content: string;
  imageUrl?: string | null;
  isAnonymous: boolean;
  username?: string | null;
  avatarUrl?: string | null;
  createdAt: Date;
}