# lib/api-zod/src/generated/types/replyRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Anonymous Messaging Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ReplyRequest {
  /**
   * @minLength 1
   * @maxLength 1000
   */
  body: string;
  imageUrl?: string | null;
}