# lib/api-zod/src/generated/types/updateAvatarRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Anonymous Messaging Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateAvatarRequest {
  avatarUrl: string;
}